+++ src/PhoneInput/PhoneInput.tsx
@@ -14,6 +14,7 @@ interface PhoneInputProps {
   onCountryCodeChange: (code: string) => void
   placeholder?: string
   label?: string | ReactNode
+  error?: boolean
   errorMessage?: string
   required?: boolean
   className?: string
@@ -25,6 +26,7 @@ const PhoneInput: React.FC<PhoneInputProps> = ({
   onCountryCodeChange,
   placeholder = 'Enter your phone number',
   label = 'Phone Number',
+  error,
   errorMessage = 'Invalid phone number',
   required = false,
   name,
@@ -59,7 +61,7 @@ const PhoneInput: React.FC<PhoneInputProps> = ({
           {required && <span className="asterick">*</span>}
         </label>
       )}
-      <div className={clsx(styles.inputContainer, { [styles.invalid]: !isValid })}>
+      <div className={clsx(styles.inputContainer, { [styles.invalid]: !isValid, [styles.inputError]: error })}>
         <select value={selectedCountryCode} onChange={handleCountryCodeChange} className={styles.countryCode}>
           {countryCodes.map((country) => (
             <option key={country.code} value={country.code}>
@@ -76,7 +78,7 @@ const PhoneInput: React.FC<PhoneInputProps> = ({
           className={styles.input}
         />
       </div>
-      {!isValid && <p className={styles.error}>{errorMessage}</p>}
+      {!isValid || (error && <p className={styles.error}>{errorMessage}</p>)}
     </div>
   )
 }